micropython_macos/requirements.txt
micropython-xmltok
